Make SRFI 132 tests portable
This was a big effort.

At the start were some of Olin Shivers' tests copied from SRFI 32.
Those won't work with SRFI 132 unless we have access to the internal
procedures in the SRFI's sample implementation, so I took them out.

The rest of the file was Will Clinger's test cases written
specifically for SRFI 132. Those were amenable to a semi-automated
conversion to SRFI 64 test cases.